requirements.txt
torch>=2.0
torchvision
transformers>=4.40
accelerate

numpy
pillow
opencv-python

scikit-learn
tqdm
pyyaml

fastapi
uvicorn
python-multipart
requests

